EVE-341 - Add code coverage reporting step
Add a buildbot step named `PublishCoverageReport`.
This step will upload all given code coverage files from the buildbot
worker to the buildbot master before to publish them to a external
code coverage service.

Actually only `codecov.io` service is supported.

Mandatory arguments:

* **repository**: Name of the git repository.
* **revision**:  The destination commit sha for the report.
* **filepaths**: List of code coverage report file path (actually only
    the first file will be published to `codecov.io`, the others will
    be ignored).

Extra arguments:

* **branch**: Name of the branch of the git repository.
* **uploadName**: Upload identifier (used for `codecov.io`).
* **flags**: List of report flag (used for `codecov.io`).
* **skipMissingFile**: Skip or no all not found report files.

Arguments used in internal to upload the reports from the buildbot
worker to the buildbot master:

* **maxSize**: Upload max size.
* **blockSize**: Upload block size.

To enable the publisment of the code coverage in Eve, we need to
define the `CODECOV_IO_UPLOAD_TOKEN` environment variable.  This
variable is the UUID token used to identify the project in `codecov.io`.
Without this assigment, the step will be skipped in buildbot.

The automatic test, executed by the CI on each commit, will use a mock
HTTP server of `codecov.io`.
To run it:

```
tox -e test -- -v tests/test_end2end.py::TestPublishCodeCoverage
```

This test will verify all the code is functionnal and the behaviour of
the publication is conform with the API of `codecov.io` (see
https://docs.codecov.io/v4.3.0/reference#upload).

This test will not verify the mocked API is always correct with the
actual `codecov.io` service.
To use the real `codecov.io` service, define the
`CODECOV_IO_UPLOAD_TOKEN` with the UUID token for the repository used
in *tests/yaml/generate_coverage_report/main.yml*.

```
CODECOV_IO_UPLOAD_TOKEN=XXXX tox -e test -- \
  -v tests/test_end2end.py::TestPublishCodeCoverage
```

For now, we use the temporary repository **scality/test_codecov_io**.
We should use directly the repository **scality/eve** to avoid to have
a repository just for this test. But to do it, we need to upgrade our
account in `codecov.io` to have more than one private repository.