EVE-390 Add StepPatcher feature
This feature adds a step that loads a yaml patch file located at
`PATCHER_FILE_PATH` and sets a `step_patcher_config` build property with its
contents. If the file is absent, the feature is disabled.

During the Step extraction, feed a StepPatcher with the config read earlier and
use it to patch the arguments given to the step creation.

Skipping steps is the only currently supported feature:

    skip_steps:
        - .*PLIP   # Skip all tests containing 'PLIP'
        - plop    # Skip all tests starting with 'plop'

A single string can be passed to this mapping

    skip_tests: ".*PLIP"