EVE-762 configurable build status filter
Replacing the hard-coded list of stages that publish
build statuses in favor of a configurable system;
The setup can now specify a list of stages that will be
notified to the githost, and another list that configures
the list of stages notified to Hipchat.

By default all stages are notified (empty setting).

Also deprecate the setting IS_PRODUCTION.
Also remove unused GITHUB_CONTEXT_STATUS.