EVE-891 Kube pod worker waits for all containers
The kube pod worker now waits for the pod to be "Running" and then for
all container inside the pod to be ready.
This is done by polling periodically kubernetes' API.
It's also as the good side effect of giving way better error feedback to
the user when the pod/containers fail to start properly.