EVE-1021 eve_api_client: Extract provider handling
Provider auth is now handled through an Factory pattern, which provides an
object based on the AbstractClass + ConcreteImplementation pattern. This allows
writing simple code to retrieve a token, only providing all possible auth
parameters (set or not).

Note that this commit bumps the Dockerfile and required pythons to 3.3 minimum,
as the methods used in githost.py require at least this version.