EVE-970 Ensure that GitHostMock checks work on dev setup & CI
The tests were written to work using the git version embedded in eve's
container image, and the CI's container image.
This led to a discrepancy between dev-run tests, and CI-run tests.

This change ensures that both older and newer versions of git are supported for
the GitHostMock tests, and that there is no need to add complexity to the mix
just to handle that smoothly.